refactor(a11y): move SSR title loaders to components/decisions/serverRenderData
lib/ is for cross-cutting infra (i18n); decision-domain helpers live in
components/decisions/*.ts (proposalContentUtils, proposalTemplate, etc.).
Colocates the cache() render loaders next to getProposalDisplayTitle.

format
