+++ apps/app/src/components/decisions/serverRenderData.ts
@@ -1,5 +1,4 @@
 import 'server-only';
-
 import { createServerUtils } from '@op/api/server';
 import { cache } from 'react';
